Added a quick hack to substituting values from a color.xml file, allowing for double-indirection.
Used argparse to parse arguments, with --color-xml-file and --viewbox-only arguments.

Added a color_map dictionary; also added associated read_color_xml() and get_color() functions that write into and read from it.

This is a hack because color_map should probably not be a module-level variable and should instead either be passed around explicitly or it and the related functions should be moved into a new class.

Added a print statement so that it is possible to know what files are being converted, particularly if using something like *.xml on the command line.

Caught any exceptions when converting individual files and printed the stack traceback.